move repositoryCard file to componentes bh folder
